bedrock/pyproject.toml
[project]
name = "bedrock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic[bedrock]>=0.52.2",
    "boto3>=1.38.32",
    "python-dotenv>=1.1.0",
]
